Speed up the engine comparison by removing redundant positions
If a position is a rotation or a reflection of another position, it
makes no sense to use it for eval coefficient comparison, since it's
identical to other position in the test set. Deduplication helps speed
up the comparison by a factor of approximately 4.26.

The deduplication code takes a page from the opening book
implementation, hence the corresponding rotation and reflection
functions have been made public.