Update reversi-tools to v1.0.0
The main change in reversi-tools v 1.0.0 is switching to the
RichPosition struct, which contains the flips that have lead to the
position and the previous move. it may help to implement incremental
evaluation (useless in this shape; may help when implementing
incremental tree traversal) and incremental Zobrist hash updating for
the puspose of low-overhead caching.

Both ventures may end up fruitless, though, making the performance
overhead counterproductive. Therefore, this change is being kept in the
branch for the time being without merging.